Do not create relay futures if they cannot be registered
Previously register_relay_future could just not register a
RelayFuture, but it was still spawned.
Upon dropping the RelayFuture, the assert in unregister_relay_future would fail.
Now we either register the RelayFuture in a way that
the destroy cleanup mechanism will not miss it,
or we just don't create it in the first place

fixes #3